requirements.txt
pytest==8.3.4
black==24.10.0
ruff==0.8.4
pylint==3.3.2
flake8==7.1.1
mypy==1.13.0
pytest-cov==6.0.0
torch==2.5.1
pandas==2.2.3
matplotlib==3.10.0
librosa==0.10.2.post1
PySoundFile==0.9.0.post1
scikit-learn==1.6.0
pillow==11.1.0
seaborn==0.13.2
opencv-python==4.10.0.84
soundfile==0.13.0
numpy==2.0.2
torchaudio==2.5.1
torchvision==0.20.1
colabcode==0.1.1
scipy==1.15.0
tqdm==4.67.1
complexipy==1.2.0
sounddevice==0.4.1